extra white space removed & flask test client moved to conftest.py
