+++ tests/test_data_response.py
@@ -1,17 +1,5 @@
 import pytest
-from app.create_app import create_app
 from flask import url_for
-
-
-@pytest.fixture()
-def flask_test_client():
-    """
-    Creates the test client we will be using to test the responses
-    from our app, this is a test fixture.
-    :return: A flask test client.
-    """
-    with create_app().test_client() as test_client:
-        yield test_client
 
 
 @pytest.mark.usefixtures("live_server")
